Dedicated GeneratorSynchronousModelGroovyExtension.groovy
